+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- macOS window can be dragged via a top title-bar strip when using `hiddenInset` title bar ([#5](https://github.com/CodingTumbleweed/markz/issues/5))
+
 ---
 
 ## [0.2.0-beta.2] -- 2026-07-27